Move RelatedLookupInfo to separate file
